Updated VMware documentation and pictures

+++ installer/scripts/genRubyInstaller.sh
@@ -1,9 +1,20 @@
 #! /bin/bash
 
-set -e
+# This script will generate a ruby.data file based on SOURCE_DIR 
+# Ruby content. Ruby.data file is consumed
+# by InstallBuilder, to enject Ruby to a native RPM 
+# or DEB installation package during the build/make process.
+#
+# It is necessary to regenerate this file if new packages have been
+# installed into Ruby, and those new packages should be deployed
+# on customer systems during OMS installation.
+# The output of this script that contains new ruby.data will be 
+# serialized in OUTPUT_RESULTS.
 
 # For running this script, you must be running on a 64-bit system!
 # This is for text substitution that happens below.
+
+set -e
 
 if [ `uname -m` != "x86_64" ]; then
     echo "This script must be run from a 64-bit system" >& 2
@@ -28,29 +39,30 @@ substitute_arch()
     echo -n "$RESULT"
 }
 
-# This script will generate a "starting point" for an InstallBuilder data file.
-# For example, this generated ruby.data in the omsagent project. This is useful
-# for a package that has an install process (i.e. make install), but we need to
-# generate an install data file for it.
-
-OUTPUT_RESULTS="/tmp/installBuilder-$USER"
+OUTPUT_RESULTS="/tmp/installBuilder-ruby-data-$USER"
 rm -f $OUTPUT_RESULTS
 
 # Evaluate any wildcards in the filename ...
-SOURCE_DIR=`stat -c "%n" ~/dev/work/oms/intermediate/*/ruby`
+SOURCE_DIR=`stat -c "%n" ../intermediate/*/100/ruby`
+
+# Verify validity of source directory
+
+if [ -z "$SOURCE_DIR" -o ! -d $SOURCE_DIR ]; then
+    echo "Intermediate SSL_100 ruby version does not exist" >& 2
+    echo "Please build omsagent configured with --enable-ulinux" >& 2
+    exit 1
+fi
+
+if [ ! -f configure -o ! -d ../build ]; then
+    echo "Must be in omsagent build directory for this script to run"
+    exit 1
+fi
 
 OUTPUT_DIR=/tmp/outdir.txt.$$
 OUTPUT_FILE=/tmp/outfile.txt.$$
 
 echo "%Directories" > $OUTPUT_DIR
 echo "%Files" > $OUTPUT_FILE
-
-# Verify validity of source directory
-
-if [ -z "$SOURCE_DIR" -o ! -d $SOURCE_DIR ]; then
-    echo "Source directory ($SOURCE_DIR) does not exist" >& 2
-    exit 1
-fi
 
 # For each destination file, generate the appropriate lines for InstallBuilder
 
@@ -59,7 +71,7 @@ for i in `find $SOURCE_DIR -name \* -print`; do
     if [ -d $i ]; then
         DIR_NAME=`echo $i | sed "s~$SOURCE_DIR~~"`
 	[ -n "$DIR_NAME" ] && DIR_NAME=`substitute_arch $DIR_NAME`
-        STAT_INFO=`stat -c "%a; %U; %G" $i`
+        STAT_INFO=`stat -c "%a; root; root" $i`
         printf "%-55s %s\n" "\${{RUBY_DEST}}${DIR_NAME};" "$STAT_INFO" >> $OUTPUT_DIR
     else
         OLD_BASE_DIR=`dirname $i`
@@ -68,12 +80,12 @@ for i in `find $SOURCE_DIR -name \* -print`; do
 
         FILE_NAME=`echo $i | sed "s~$SOURCE_DIR~~"`
 	FILE_NAME=`substitute_arch $FILE_NAME`
-        STAT_INFO=`stat -c "%a; %U; %G" $i`
+        STAT_INFO=`stat -c "%a; root; root" $i`
         printf "%-72s %-64s %s\n" "\${{RUBY_DEST}}${FILE_NAME};" "\${{RUBY_INT}}${FILE_NAME};" "$STAT_INFO" >> $OUTPUT_FILE
     fi
 done
 
-# Generate results
+# Generate ruby.data
 
 echo "Writing results to file: $OUTPUT_RESULTS"
